renamed SectionContentNotification to message
