+++ src/components/section/Section/ConvenientSectionContentMessage.tsx
@@ -3,19 +3,15 @@ import IconList from '@aboutbits/react-material-icons/dist/IconList'
 import IconWarning from '@aboutbits/react-material-icons/dist/IconWarning'
 import { Tone } from '../../types'
 import {
-  SectionContentNotification,
+  SectionContentMessage,
   SectionContentNotificationProps,
-} from './SectionContentNotification'
+} from './SectionContentMessage'
 
 export function SectionContentError({
   ...props
 }: Omit<SectionContentNotificationProps, 'tone' | 'icon'>): ReactElement {
   return (
-    <SectionContentNotification
-      tone={Tone.critical}
-      icon={IconWarning}
-      {...props}
-    />
+    <SectionContentMessage tone={Tone.critical} icon={IconWarning} {...props} />
   )
 }
 
@@ -23,16 +19,6 @@ export function SectionContentEmpty({
   ...props
 }: Omit<SectionContentNotificationProps, 'tone' | 'icon'>): ReactElement {
   return (
-    <SectionContentNotification
-      tone={Tone.neutral}
-      icon={IconList}
-      {...props}
-    />
+    <SectionContentMessage tone={Tone.neutral} icon={IconList} {...props} />
   )
-}
-
-export function SectionContentMessage({
-  ...props
-}: Omit<SectionContentNotificationProps, 'tone' | 'icon'>): ReactElement {
-  return <SectionContentNotification tone={Tone.neutral} {...props} />
 }
