+++ src/components/section/Section/SectionContentMessage.stories.mdx
@@ -1,23 +1,20 @@
 import { Meta, Story, Canvas, ArgsTable } from '@storybook/addon-docs'
 import IconList from '@aboutbits/react-material-icons/dist/IconList'
 import { Tone } from '../../types'
-import { SectionContentNotification } from './SectionContentNotification'
+import { SectionContentMessage } from './SectionContentMessage'
 import {
   SectionContentEmpty,
   SectionContentError,
-  SectionContentMessage,
-} from './ConvenientSectionContentNotification'
+} from './ConvenientSectionContentMessage'
 
 <Meta
-  title="Components/Section/SectionContentNotification"
-  component={SectionContentNotification}
+  title="Components/Section/SectionContentMessage"
+  component={SectionContentMessage}
 />
 
-# SectionContentNotification
+# SectionContentMessage
 
-export const Template = ({ ...args }) => (
-  <SectionContentNotification {...args} />
-)
+export const Template = ({ ...args }) => <SectionContentMessage {...args} />
 
 <Canvas>
   <Story
