+++ src/examples/list.stories.mdx
@@ -33,7 +33,7 @@ import {
 import { SectionFooterWithPaginationInMemory } from '../components/pagination'
 import { Select, Variant as InputVariant } from '../components/form'
 import { Tone } from '../components/types'
-import { SectionContentNotification } from '../components/section/Section/SectionContentNotification'
+import { SectionContentMessage } from '../components/section/Section/SectionContentMessage'
 
 <Meta title="Examples/List" />
 
@@ -56,7 +56,7 @@ export const TemplateDefault = ({ numberOfTotalItems }) => {
       </SectionHeader>
       <SectionContainer>
         {content.length === 0 ? (
-          <SectionContentNotification
+          <SectionContentMessage
             tone={Tone.informative}
             text="No users found in database."
           />
@@ -182,7 +182,7 @@ export const TemplateComplexList = ({
       </SectionHeader>
       <SectionContainer>
         {content.length === 0 ? (
-          <SectionContentNotification
+          <SectionContentMessage
             tone={Tone.informative}
             text="No users found in database."
           />
@@ -263,7 +263,7 @@ export const TemplateSearch = ({ numberOfTotalItems }) => {
       </SectionFilter>
       <SectionContainer>
         {content.length === 0 ? (
-          <SectionContentNotification
+          <SectionContentMessage
             tone={Tone.informative}
             text="No users found in database."
           />
@@ -479,7 +479,7 @@ export const TemplateListWithFilterDialog = ({
         </SectionHeader>
         <SectionContainer>
           {content.length === 0 ? (
-            <SectionContentNotification
+            <SectionContentMessage
               tone={Tone.informative}
               text="No users found in database."
             />
